Modified tranquiliser tests and tranquiliser program (changed variable names), added to axelrod_second and test_axelrod_second and deleted standalone tranquiliser.py and test_tranquiliser.py files
