+++ axelrod/tests/strategies/test_axelrod_second.py
@@ -5,6 +5,8 @@ import random
 import axelrod
 from .test_player import TestPlayer
 
+from axelrod.interaction_utils import compute_final_score
+
 C, D = axelrod.Action.C, axelrod.Action.D
 
 
@@ -159,21 +161,10 @@ class TestGladstein(TestPlayer):
         self.versus_test(opponent, expected_actions=actions,
                          attrs={'patsy': False})
 
-import axelrod
-from Axelrod.axelrod.tests.strategies.test_player import TestPlayer
-import unittest
-
-C, D = axelrod.Action.C, axelrod.Action.D
-
 
 class TestTranquiliser(TestPlayer):
-    """
- Note that this test is referred to in the documentation as an example on
- writing tests.  If you modify the tests here please also modify the
- documentation.
- """
 
-    name = "Tit For Tat"
+    name = "Tranquiliser"
     player = axelrod.Tranquiliser
     expected_classifier = {
         'memory_depth': float('inf'),
@@ -185,5 +176,94 @@ class TestTranquiliser(TestPlayer):
         'manipulates_state': False
     }
 
-    def test_strategy(self):
-        self.assertEqual(player.ratioFD2, 0)
+
+    # test for initalised variables
+
+    def test_init(self):
+
+        player = axelrod.Tranquiliser()
+
+        self.assertEqual(player.P, 1.1)
+        self.assertEqual(player.FD, 0)
+        self.assertEqual(player.consecutive_defections, 0)
+        self.assertEqual(player.ratio_FD1, 5)
+        self.assertEqual(player.ratio_FD2, 0)
+        self.assertEqual(player.ratio_FD1_count, 1)
+        self.assertEqual(player.ratio_FD2_count, 1)
+        self.assertEqual(player.score, None)
+        self.current_score = 0
+
+    def test_score_bad(self):
+
+        # Tests whether TitForTat is played given score is below 1.75
+
+        opponent = axelrod.Defector()
+        actions = [(C, D)] + [(D, D)] * 20
+        self.versus_test(opponent, expected_actions=actions, attrs={"P":1.1, "FD":0, "consecutive_defections":20, "ratio_FD1":5, 
+                                                                    "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})
+
+        opponent = axelrod.MockPlayer([C] * 2 + [D] * 9 + [C] * 4 )
+        actions = [(C, C)] + [(C, C)] + [(C, D)] * 2 + [(D, D)] * 7 + [(D, C)] + [(C, C)] * 3  
+        self.versus_test(opponent = opponent, expected_actions=actions, attrs={"P":1.8667, "FD":0, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})
+
+        # If score is between 1.75 and 2.25, may cooperate or defect
+        opponent = axelrod.MockPlayer(actions=[D] * 8 + [C] * 5 + [D])
+        actions = [(C, D)] + [(D, D)] * 7 + [(D, C)] + [(C, C)] * 4
+        # average_score_per_turn = 1.875
+        actions += ([(C, D)]) # <-- Random
+        self.versus_test(opponent, expected_actions=actions, seed=1, attrs={"P": 0.9203, "FD":0, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})
+
+        opponent = axelrod.MockPlayer(actions=[D] * 8 + [C] * 4 + [D])
+        actions = [(C, D)] + [(D, D)] * 7 + [(D, C)] + [(C, C)] * 3         
+        # average_score_per_turn = 1.875
+        actions += [(D, D)] # <-- Random
+        self.versus_test(opponent, expected_actions=actions, seed=10, attrs={"P": 0.891, "FD":0, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})
+
+        # If score is greater than 2.25 either cooperate or defect, if turn number <= 4; cooperate
+        
+        opponent = axelrod.MockPlayer(actions=[C] * 5)
+        actions = [(C, C)] * 4 + [(C, C)]
+        self.versus_test(opponent, expected_actions=actions, seed=1, attrs={"P": 0.99, "FD":0, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})
+
+        opponent = axelrod.MockPlayer(actions=[C] * 5)
+        actions = [(C, C)] * 4 + [(D, C)]
+        self.versus_test(opponent, expected_actions=actions, seed=70,  attrs={"P": 0.99, "FD":1, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})
+
+        # Given score per turn is greater than 2.25, Tranquiliser will never defect twice in a row
+        
+        opponent = axelrod.MockPlayer(actions = [C] * 5)
+        actions = [(C, C)] * 4 + [(D, C)]
+        self.versus_test(opponent, expected_actions=actions, seed=70, attrs={"P": 0.99, "FD":1, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})
+        # Tests defection probability if score == good
+
+        opponent = axelrod.MockPlayer(actions=[C] * 5)
+        actions = [(C, C)] * 4 + [(D, C)]
+        self.versus_test(opponent, expected_actions=actions, seed=70, attrs={"P": 0.99, "FD":1, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1})        
+       
+       # Ensures FD1 values are calculated
+
+        opponent = axelrod.MockPlayer(actions=[C] * 6)
+        actions = [(C, C)] * 4 + [(D, C)] + [(C, C)]
+        self.versus_test(opponent, expected_actions=actions, seed=70, attrs={"P": 0.99, "FD":2, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":0, "ratio_FD1_count":2, "ratio_FD2_count":1})
+        
+        # Ensures FD2 values are calculated
+        
+        opponent = axelrod.MockPlayer(actions=[C] * 6)
+        actions = [(C, C)] * 4 + [(D, C)] + [(C, C)] + [(C, C)]
+        self.versus_test(opponent, expected_actions=actions, seed=70,  attrs={"P": 0.99, "FD":0, "consecutive_defections":0, "ratio_FD1":5, 
+                                                                                "ratio_FD2":1.5, "ratio_FD1_count":2, "ratio_FD2_count":2})
+        
+        # Ensures scores are being counted
+
+        opponent = axelrod.Defector()
+        actions = [(C, D)] + [(D, D)] * 19
+        self.versus_test(opponent, expected_actions=actions, attrs={"P": 1.1, "FD":0, "consecutive_defections":19, "ratio_FD1":5, 
+                                                                    "ratio_FD2":0, "ratio_FD1_count":1, "ratio_FD2_count":1}) # Check
